@@ -1,6 +1,6 @@
 {
 	"name": "picture_slicer_multi",
-	"version": "0.0.2",
+	"version": "0.0.3",
 	"scripts": {
 		"dev": "vite dev --host 0.0.0.0",
 		"build": "vite build",
@@ -13,19 +13,19 @@
 	"devDependencies": {
 		"@sveltejs/adapter-auto": "^3.2.2",
 		"@sveltejs/adapter-static": "^3.0.2",
-		"@sveltejs/kit": "^2.5.20",
-		"@sveltejs/vite-plugin-svelte": "^3.1.1",
-		"@typescript-eslint/eslint-plugin": "^5.62.0",
-		"@typescript-eslint/parser": "^5.62.0",
+		"@sveltejs/kit": "^2.8.0",
+		"@sveltejs/vite-plugin-svelte": "^4.0.0",
+		"@types/eslint": "^9.6.0",
 		"eslint": "^8.57.0",
-		"eslint-config-prettier": "^8.10.0",
+		"eslint-config-prettier": "^9.1.0",
 		"eslint-plugin-svelte": "^2.43.0",
-		"prettier": "^2.8.8",
-		"prettier-plugin-svelte": "^2.10.1",
-		"svelte": "^4.2.18",
-		"svelte-check": "^3.8.5",
-		"tslib": "^2.6.3",
-		"typescript": "^5.5.4",
+		"globals": "^15.0.0",
+		"prettier": "^3.3.2",
+		"prettier-plugin-svelte": "^3.2.6",
+		"svelte": "^5.1.12",
+		"svelte-check": "^4.0.5",
+		"typescript": "^5.6.3",
+		"typescript-eslint": "^8.13.0",
 		"vite": "^5.3.5"
 	},
 	"type": "module",
