@@ -1,8 +1,13 @@
 <script lang="ts">
+	interface Props {
+		children?: import('svelte').Snippet;
+	}
+
+	let { children }: Props = $props();
 </script>
 
 <section>
-	<slot />
+	{@render children?.()}
 </section>
 
 <style>
