@@ -1,8 +1,13 @@
 <script lang="ts">
+	interface Props {
+		children?: import('svelte').Snippet;
+	}
+
+	let { children }: Props = $props();
 </script>
 
 <svelte:head>
 	<title>Picture slicer</title>
 </svelte:head>
 
-<slot />
+{@render children?.()}
